improve CLI performance
